requirements.txt
Django==4.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-silk==5.0.2
redis==4.4.0
hiredis==2.1.0
django-filter==22.1
python-decouple==3.6
